refactor: :recycle: pet profile refactor
moved pet profile components from profile page into standalone components
